qa_gen_pipelines/requirements.txt
# Core dependencies
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3

# LLM and RAG
lightrag>=1.3.0
openai>=1.0.0
anthropic>=0.8.0

# Text processing
nltk>=3.8.1
tiktoken>=0.5.0
markdown>=3.5.0

# Data handling
jsonlines>=4.0.0
pydantic>=2.5.0
PyYAML>=6.0.1

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
loguru>=0.7.0

# System dependencies
numpy>=1.24.0
pandas>=2.0.0

# Optional: For GPU acceleration
# torch>=2.0.0
# 我们将不再使用GPU版本，而是明确指定CPU版本以确保成功